Refactor: resetPw 기능 findPw 파일로 이동 - isValidResponse, isErrorResponse 사용하게 수정 - 주석 추가
